Refactored and improved performance `HomeView`'s table view and collection view data sources.

+++ netflix/Domain/Entities/Media.swift
@@ -15,7 +15,7 @@ struct Media {
     
     var id: String?
     var title: String
-    var rating: CGFloat
+    var rating: Float
     var description: String
     var cast: String
     var isHD: Bool
